Permit passing download options from tswrap
Fix the case where the tagsoup wrapper script has been extended with
custom filters in order to ensure that if $TAGSOUP_DLOPTS has been set
in the filter then tagsoup runs with $DLOPTS set to the same value (and
then to use it verbatim unless explicitly overridden).

This ensures the wrapper script can retrieve from pages that blacklist
wget as a user agent, or that expect the optional HTTP referer to be set
when retrieving images used in a given page.